Add SPAWN, JOIN and YIELD bytecode
